+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 -   Added `@author` to the default list of recognized tags, #2603.
 -   Anchor links are no longer incorrectly checked for relative paths, #2604.
 -   Fixed an issue where line numbers reported in error messages could be incorrect, #2605.
+-   Fixed relative link detection for markdown links containing code in their label, #2606.
 
 ### Thanks!